refactor(RegisterScreen): mover estilos para arquivo styles.js
