+++ mavis/src/RegisterScreen/styles.js
@@ -1,0 +1,33 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    padding: 20,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fff',
+  },
+  title: {
+    textAlign: 'center',
+    marginBottom: 20,
+  },
+  input: {
+    marginBottom: 6,
+    width: '100%',
+  },
+  button: {
+    marginTop: 10,
+    alignSelf: 'center',
+    width: '100%',
+  },
+  registerText: {
+    textAlign: 'center',
+    marginTop: 16,
+  },
+  link: {
+    color: '#007bff',
+    textDecorationLine: 'underline',
+  }
+});
+
+export default styles
